+++ src/components/MovieHeader/MovieHeader.css
@@ -1,0 +1,89 @@
+header {
+  width: 100%;
+  height: 72px;
+  background-color: #222;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 20px;
+  border-bottom: 1px solid white;
+  margin-bottom: 48px;
+}
+
+header h1 {
+  cursor: pointer;
+  user-select: none;
+  font-size: 2rem;
+  font-weight: bold;
+  letter-spacing: -0.1rem;
+  color: #f33f3f;
+}
+
+header > .search-box {
+  background: #fff;
+  padding: 8px;
+  border-radius: 4px;
+}
+
+header .search-box > input {
+  border: 0;
+}
+
+header .search-box > .search-button {
+  width: 14px;
+  border: 0;
+  text-indent: -1000rem;
+  background: url('../../assets/image/search_button.png') transparent no-repeat 0 1px;
+  background-size: contain;
+}
+
+.search-close-button {
+  display: none;
+  width: 30px;
+  height: 33.6px;
+  font-size: 16px;
+  border-radius: 4px;
+  border: 0;
+  background-color: #fff;
+}
+
+.search-close-button:hover {
+  background-color: #ffffffd7;
+}
+
+@media (max-width: 750px) {
+  .search-box:hover {
+    background-color: #ffffffd7;
+  }
+
+  .search-box > input {
+    width: 0;
+    padding: 0;
+  }
+
+  header.clicked > h1 {
+    display: none;
+  }
+
+  header.clicked input {
+    animation: expand 1s forwards;
+  }
+
+  header.clicked .search-box {
+    background-color: #fff;
+  }
+
+  header.clicked .search-close-button {
+    display: block;
+  }
+
+  @keyframes expand {
+    from {
+      width: 0;
+    }
+
+    to {
+      width: 70vw;
+    }
+  }
+}
