+++ src/components/MovieItem/MovieItemDetailTitleContainer/MovieItemDetailTitleContainer.css
@@ -1,0 +1,30 @@
+.movie-item-detail_title-container {
+  border-bottom: 1px solid rgba(241, 241, 241, 0.25);
+}
+
+.movie-item-detail_title {
+  color: #fff;
+  font-size: 24px;
+  text-align: center;
+  margin: 18px 32px;
+}
+
+.movie-item-detail_close-button {
+  width: 36px;
+  height: 36px;
+  position: absolute;
+  top: 12px;
+  right: 32px;
+  border-radius: 100px;
+  font-size: 20px;
+  background-color: rgba(56, 56, 57, 1);
+  color: rgba(241, 241, 241, 1);
+  padding: 0;
+  border: 0;
+}
+
+@media (max-width: 750px) {
+  .movie-item-detail_title {
+    text-align: left;
+  }
+}
