+++ src/components/MovieItem/MyRatingContainer/MyRatingContainer.css
@@ -1,0 +1,56 @@
+.movie-item-detail_my-rating-container {
+  width: 438px;
+  height: 64px;
+  border-radius: 16px;
+  background-color: rgba(56, 56, 57, 1);
+}
+
+.movie-item-detail_my-rating-content {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  margin: 16px 0 16px 16px;
+  font-size: 16px;
+}
+
+.movie-item-detail_my-rating-content > h2 {
+  padding-top: 2px;
+  font-weight: bold;
+}
+
+.movie-item-detail_my-rating-content > p {
+  padding-top: 5px;
+}
+
+.movie-item-detail_my-rating-stars {
+  display: flex;
+  height: 32px;
+}
+
+.movie-item-detail_my-rating-star {
+  cursor: pointer;
+  width: 32px;
+}
+
+@media (max-width: 1000px) {
+  .movie-item-detail_my-rating-container {
+    width: 384px;
+    height: 64px;
+  }
+}
+
+@media (max-width: 750px) {
+  .movie-item-detail_my-rating-container {
+    width: 100%;
+    height: 64px;
+  }
+
+  .movie-item-detail_my-rating-content {
+    width: 250px;
+    margin: 16px auto;
+  }
+
+  .movie-item-detail_my-rating-content > p:last-child {
+    display: none;
+  }
+}
