+++ src/components/StarImg/StarImg.ts
