+++ src/components/ToastPopup/ToastPopup.css
@@ -1,0 +1,18 @@
+#toast_message {
+  opacity: 0;
+  position: fixed;
+  top: 80px;
+  right: -100px;
+  transform: translate(0, -50%);
+  padding: 10px 50px;
+  background: #f33f3f;
+  border-radius: 100px 0 0 100px;
+  color: #fff;
+  transition: all 0.5s;
+}
+
+#toast_message.active {
+  opacity: 100%;
+  top: 80px;
+  right: 0px;
+}
