+++ src/css/common.css
@@ -31,68 +31,6 @@ button {
 *:focus {
   outline: none;
 }
-.item-view,
-.item-test {
-  width: 100%;
-}
-
-.item-view {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 912px;
-  margin: 0 auto;
-}
-
-.item-view h2 {
-  font-size: 2rem;
-  font-weight: bold;
-  user-select: none;
-}
-
-/* https://andrew.hedges.name/experiments/aspect_ratio/ */
-
-.item-list {
-  display: grid;
-  grid-template-columns: repeat(4, 180px);
-  grid-column-gap: 64px;
-  grid-row-gap: 64px;
-  margin: 32px 0;
-}
-
-.item-card {
-  display: flex;
-  flex-direction: column;
-  cursor: pointer;
-}
-
-.item-thumbnail {
-  border-radius: 8px;
-  width: 180px;
-  height: 270px;
-  background-size: contain;
-}
-
-.item-title {
-  margin-top: 16px;
-  font-size: 1.2rem;
-  font-weight: bold;
-}
-
-.item-score {
-  margin-top: 16px;
-  font-size: 1.2rem;
-}
-
-.item-score::after {
-  margin-left: 8px;
-}
-
-.item-title.skeleton::after,
-.item-score.skeleton::after {
-  font-size: 0;
-  content: 'loading';
-}
 
 .full-width {
   width: 100%;
@@ -100,388 +38,4 @@ button {
 
 .last-item {
   margin-top: 48px;
-}
-
-button.btn {
-  border: 0;
-  border-radius: 8px;
-  height: 30px;
-  color: #fff;
-}
-
-button.primary {
-  background: #f33f3f;
-}
-
-.item-card .skeleton {
-  background: linear-gradient(-90deg, #aaa, #f0f0f0, #aaa, #f0f0f0);
-  background-size: 400%;
-  animation: skeleton-animation 5s infinite ease-out;
-  border-radius: 8px;
-}
-
-@keyframes skeleton-animation {
-  0% {
-    background-position: 0% 50%;
-  }
-  50% {
-    background-position: 100% 50%;
-  }
-  100% {
-    background-position: 0% 50%;
-  }
-}
-header {
-  width: 100%;
-  height: 72px;
-  background-color: #222;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 20px;
-  border-bottom: 1px solid white;
-  margin-bottom: 48px;
-}
-
-header h1 {
-  cursor: pointer;
-  user-select: none;
-  font-size: 2rem;
-  font-weight: bold;
-  letter-spacing: -0.1rem;
-  color: #f33f3f;
-}
-
-header > .search-box {
-  background: #fff;
-  padding: 8px;
-  border-radius: 4px;
-}
-
-header .search-box > input {
-  border: 0;
-}
-
-header .search-box > .search-button {
-  width: 14px;
-  border: 0;
-  text-indent: -1000rem;
-  background: url('../assets/image/search_button.png') transparent no-repeat 0 1px;
-  background-size: contain;
-}
-
-.search-close-button {
-  display: none;
-  width: 30px;
-  height: 33.6px;
-  font-size: 16px;
-  border-radius: 4px;
-  border: 0;
-  background-color: #fff;
-}
-
-.search-close-button:hover {
-  background-color: #ffffffd7;
-}
-
-#search-mini-box {
-  display: none;
-}
-
-.open {
-  display: block;
-}
-
-.modal {
-  display: none;
-}
-
-.modal--open {
-  display: block;
-}
-
-.modal-backdrop {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-
-  background: rgba(0, 0, 0, 0.35);
-}
-
-.movie-item-detail-container {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 826px;
-  height: 577px;
-  gap: 0px;
-  border-radius: 8px;
-  background-color: rgba(33, 33, 34, 1);
-}
-
-.movie-item-detail_title-container {
-  border-bottom: 1px solid rgba(241, 241, 241, 0.25);
-}
-
-.movie-item-detail_title {
-  color: #fff;
-  font-size: 24px;
-  text-align: center;
-  margin: 18px 32px;
-}
-
-.movie-item-detail_close-button {
-  width: 36px;
-  height: 36px;
-  position: absolute;
-  top: 12px;
-  right: 32px;
-  border-radius: 100px;
-  font-size: 20px;
-  background-color: rgba(56, 56, 57, 1);
-  color: rgba(241, 241, 241, 1);
-  padding: 0;
-  border: 0;
-}
-
-.movie-item-detail_content-container {
-  display: flex;
-  height: 433px;
-  gap: 32px;
-  margin: 36px 32px 48px 32px;
-}
-
-.movie-item-detail_thumbnail {
-  width: 292px;
-  height: 433px;
-  top: 96px;
-  left: 32px;
-  background-size: contain;
-}
-
-.movie-item-detail_content {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 16px;
-}
-
-.movie-item-detail_content-detail {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.movie-item-detail_genres-rating-container {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-  font-size: 16px;
-}
-
-.movie-item-detail_genres-rating-container p {
-  padding-top: 5px;
-}
-
-.movie-item-detail_rating {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  margin: 0;
-}
-
-.movie-item-detail_overview {
-  display: -webkit-box;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 14;
-  overflow: hidden;
-  line-height: 24px;
-  font-size: 16px;
-}
-
-.movie-item-detail_my-rating-container {
-  width: 438px;
-  height: 64px;
-  border-radius: 16px;
-  background-color: rgba(56, 56, 57, 1);
-}
-
-.movie-item-detail_my-rating-content {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  margin: 16px 0 16px 16px;
-  font-size: 16px;
-}
-
-.movie-item-detail_my-rating-content > h2 {
-  padding-top: 2px;
-  font-weight: bold;
-}
-
-.movie-item-detail_my-rating-content > p {
-  padding-top: 5px;
-}
-
-.movie-item-detail_my-rating-stars {
-  display: flex;
-  height: 32px;
-}
-
-.movie-item-detail_my-rating-star {
-  cursor: pointer;
-  width: 32px;
-}
-
-#toast_message {
-  opacity: 0;
-  position: fixed;
-  top: 80px;
-  right: -100px;
-  transform: translate(0, -50%);
-  padding: 10px 50px;
-  background: #f33f3f;
-  border-radius: 100px 0 0 100px;
-  color: #fff;
-  transition: all 0.5s;
-}
-
-#toast_message.active {
-  opacity: 100%;
-  top: 80px;
-  right: 0px;
-}
-
-@media (max-width: 1000px) {
-  .item-view {
-    width: 674px;
-  }
-
-  .item-list {
-    grid-template-columns: repeat(3, 180px);
-    grid-row-gap: 55px;
-    margin: 27.5px 0;
-  }
-
-  .movie-item-detail-container {
-    width: 740px;
-    height: 544px;
-  }
-
-  .movie-item-detail_content-container {
-    height: 400px;
-  }
-
-  .movie-item-detail_thumbnail {
-    width: 260px;
-    height: 400px;
-  }
-
-  .movie-item-detail_overview {
-    -webkit-line-clamp: 12;
-  }
-
-  .movie-item-detail_my-rating-container {
-    width: 384px;
-    height: 64px;
-  }
-}
-
-@media (max-width: 750px) {
-  .search-box:hover {
-    background-color: #ffffffd7;
-  }
-
-  .search-box > input {
-    width: 0;
-    padding: 0;
-  }
-
-  header.clicked > h1 {
-    display: none;
-  }
-
-  header.clicked input {
-    animation: expand 1s forwards;
-  }
-
-  header.clicked .search-box {
-    background-color: #fff;
-  }
-
-  header.clicked .search-close-button {
-    display: block;
-  }
-
-  @keyframes expand {
-    from {
-      width: 0;
-    }
-
-    to {
-      width: 70vw;
-    }
-  }
-
-  .item-view {
-    width: 316px;
-  }
-
-  .item-list {
-    grid-template-columns: repeat(2, 140px);
-    grid-column-gap: 36px;
-    grid-row-gap: 32px;
-    margin: 18px 0;
-  }
-
-  .item-card {
-    height: 308px;
-  }
-
-  .item-thumbnail {
-    width: 140px;
-  }
-
-  .movie-item-detail-container {
-    width: 100%;
-    height: 485px;
-    top: 100%;
-    transform: translate(-50%, -100%);
-  }
-
-  .movie-item-detail_content {
-    width: 100%;
-  }
-
-  .movie-item-detail_content-container {
-    height: 341px;
-  }
-
-  .movie-item-detail_thumbnail {
-    display: none;
-  }
-
-  .movie-item-detail_title {
-    text-align: left;
-  }
-
-  .movie-item-detail_overview {
-    -webkit-line-clamp: 9;
-  }
-  .movie-item-detail_my-rating-container {
-    width: 100%;
-    height: 64px;
-  }
-
-  .movie-item-detail_my-rating-content {
-    width: 250px;
-    margin: 16px auto;
-  }
-
-  .movie-item-detail_my-rating-content > p:last-child {
-    display: none;
-  }
 }
